refactor erc20Balance method
