+++ web/packages/api/src/assets_v2.ts
@@ -1032,21 +1032,16 @@ export async function getAssetHubConversionPalletSwap(
     return BigInt(asset1Balance)
 }
 
-export const assetErc20Balance = async (
-    context: Context,
-    token: string,
-    owner: string
-): Promise<{
-    balance: bigint
-    gatewayAllowance: bigint
-}> => {
-    const [ethereum, gateway] = await Promise.all([context.ethereum(), context.gateway()])
-
-    const tokenContract = IERC20__factory.connect(token, ethereum)
-    const gatewayAddress = await gateway.getAddress()
+export async function erc20Balance(
+    ethereum: AbstractProvider,
+    tokenAddress: string,
+    owner: string,
+    spender: string
+) {
+    const tokenContract = IERC20__factory.connect(tokenAddress, ethereum)
     const [balance, gatewayAllowance] = await Promise.all([
         tokenContract.balanceOf(owner),
-        tokenContract.allowance(owner, gatewayAddress),
+        tokenContract.allowance(owner, spender),
     ])
     return {
         balance,
