+++ web/packages/api/src/transfers/toPolkadot/pnaToParachain.ts
@@ -13,7 +13,6 @@ import {
     dryRunAssetHub,
     dryRunDestination,
     encodeNativeAsset,
-    erc20Balance,
     hexToBytes,
     Transfer,
     validateAccount,
@@ -22,7 +21,7 @@ import {
 } from "../../toPolkadotSnowbridgeV2"
 import { accountId32Location, DOT_LOCATION, erc20Location } from "../../xcmBuilder"
 import { paraImplementation } from "../../parachains"
-import { ETHER_TOKEN_ADDRESS, swapAsset1ForAsset2 } from "../../assets_v2"
+import { erc20Balance, ETHER_TOKEN_ADDRESS, swapAsset1ForAsset2 } from "../../assets_v2"
 import { beneficiaryMultiAddress, padFeeByPercentage, paraIdToSovereignAccount } from "../../utils"
 import { FeeInfo, resolveInputs, ValidationLog, ValidationReason } from "../../toPolkadot_v2"
 import {
